refactor: extract types, utils, and StatusCard from WorkflowForm.tsx
- WorkflowFormTypes.ts: HookDraft, StatusDraft, AgentConfigDraft, genKey
- WorkflowFormUtils.ts: workflowToDrafts, buildProtoPayload (now takes params)
- StatusCard.tsx: status card rendering with internal HookRow component

WorkflowForm.tsx reduced from 861 to 385 lines.
